# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLM Regression Lab — proposed contract",
    "version": "1.0.0",
    "description": "Phase 2 blueprint. No running server is provided."
  },
  "servers": [
    {
      "url": "https://eval.example.invalid",
      "description": "Placeholder only"
    }
  ],
  "paths": {
...
    "/v1/jobs": {
      "post": {
        "operationId": "submitJob",
        "summary": "PLANNED: submitJob",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "jobs:write",
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              },
              "Location": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 256
                },
                "description": "Relative resource URL"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "202": {
            "description": "Successful response",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              },
              "Location": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 256
                },
                "description": "Relative resource URL"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Error; see error code and retryable fields",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "pattern": "^[\\x20-\\x7e]+$"
            }
          }
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "getJob",
        "summary": "PLANNED: getJob",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "jobs:read",
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Error; see error code and retryable fields",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancelJob",
        "summary": "PLANNED: cancelJob",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "jobs:write",
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              },
              "Location": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 256
                },
                "description": "Relative resource URL"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "202": {
            "description": "Successful response",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              },
              "Location": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 256
                },
                "description": "Relative resource URL"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Error; see error code and retryable fields",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/v1/jobs/{job_id}/report": {
      "get": {
        "operationId": "getReport",
        "summary": "PLANNED: getReport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "reports:read",
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "default": {
            "description": "Error; see error code and retryable fields",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/v1/jobs/{job_id}/gate": {
      "get": {
        "operationId": "getGate",
        "summary": "PLANNED: getGate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "reports:read",
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gate"
                }
              }
            }
          },
          "default": {
            "description": "Error; see error code and retryable fields",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "description": "Server request identifier"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "JobRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "1.0",
            "type": "string"
          },
          "dataset_version": {
            "type": "string",
            "minLength": 1,
            "maxLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$"
          },
          "config_version": {
            "type": "string",
            "minLength": 1,
            "maxLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$"
          },
          "baseline_job_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 36,
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "deadline_seconds": {
            "type": "integer",
            "minimum": 60,
            "maximum": 3600
          },
          "source_revision": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "pattern": "^[0-9a-f]{40}$"
          }
        },
        "required": [
          "schema_version",
          "dataset_version",
          "config_version",
          "baseline_job_id",
          "deadline_seconds",
          "source_revision"
        ]
      },
      "ExecutionError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "PROVIDER_TIMEOUT",
              "PROVIDER_RATE_LIMIT",
              "PROVIDER_AUTH",
              "ENGINE_CRASH",
              "INVALID_ENGINE_OUTPUT",
              "ARTIFACT_INTEGRITY",
              "DISPATCH_FAILED",
              "LEASE_EXPIRED",
              "DEADLINE_EXCEEDED",
              "BUDGET_EXCEEDED",
              "CANCELLED"
            ]
          },
          "retryable": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          }
        },
        "required": [
          "code",
          "retryable",
          "message"
        ]
      },
      "Job": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 36,
            "format": "uuid"
          },
          "request": {
            "$ref": "#/components/schemas/JobRequest"
          },
          "state": {
            "type": "string",
            "enum": [
              "QUEUED",
              "DISPATCHING",
              "RUNNING",
              "RETRY_WAIT",
              "CANCEL_REQUESTED",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED"
            ]
          },
          "attempt": {
            "type": "integer",
            "minimum": 0,
            "maximum": 3
          },
          "created_at": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "format": "date-time"
          },
          "deadline_at": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "format": "date-time"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionError"
              },
              {
                "type": "null"
              }
            ]
          },
          "links": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "self": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              "report": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              "gate": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              }
            },
            "required": [
              "self",
              "report",
              "gate"
            ]
          }
        },
        "required": [
          "job_id",
          "request",
          "state",
          "attempt",
          "created_at",
          "updated_at",
          "deadline_at",
          "error",
          "links"
        ]
      },
...
      "Gate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 36,
            "format": "uuid"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PASS",
              "REGRESSION",
              "INCONCLUSIVE"
            ]
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "WITHIN_LIMITS",
                "QUALITY_REGRESSION",
                "LATENCY_REGRESSION",
                "COST_REGRESSION",
                "EXECUTION_FAILED",
                "CANCELLED",
                "BASELINE_REQUIRED",
                "INCOMPATIBLE_BASELINE",
                "INSUFFICIENT_EVIDENCE",
                "MISSING_METRIC",
                "UNKNOWN_COST"
              ]
            },
            "minItems": 1,
            "maxItems": 12,
            "uniqueItems": true
          },
          "policy_version": {
            "const": "paired-v1",
            "type": "string"
          },
          "policy_digest": {
            "type": "string",
            "minLength": 1,
            "maxLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$"
          },
          "decided_at": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "format": "date-time"
          },
          "decisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MetricDecision"
            },
            "minItems": 0,
            "maxItems": 20
          }
        },
        "required": [
          "job_id",
          "verdict",
          "reasons",
          "policy_version",
          "policy_digest",
          "decided_at",
          "decisions"
        ]
      },
...
      "Report": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "1.0",
            "type": "string"
          },
          "job_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 36,
            "format": "uuid"
          },
          "dataset_version": {
            "type": "string",
            "minLength": 1,
            "maxLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$"
          },
          "config_version": {
            "type": "string",
            "minLength": 1,
            "maxLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$"
          },
          "baseline_job_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 36,
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "completed_at": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "format": "date-time"
          },
          "manifest": {
            "$ref": "#/components/schemas/Manifest"
          },
          "coverage": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "expected": {
                "type": "integer",
                "minimum": 1,
                "maximum": 1000000
              },
              "valid": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              }
            },
            "required": [
              "expected",
              "valid"
            ]
          },
          "metrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Metric"
            },
            "minItems": 1,
            "maxItems": 20
          },
          "cost": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "estimated_usd": {
                "type": "string",
                "minLength": 1,
                "maxLength": 32,
                "pattern": "^[0-9]+\\.[0-9]{6}$"
              },
              "input_tokens": {
                "type": "integer",
                "minimum": 0,
                "maximum": 10000000
              },
              "output_tokens": {
                "type": "integer",
                "minimum": 0,
                "maximum": 10000000
              },
              "usage_complete": {
                "type": "boolean"
              }
            },
            "required": [
              "estimated_usd",
              "input_tokens",
              "output_tokens",
              "usage_complete"
            ]
          },
          "gate": {
            "$ref": "#/components/schemas/Gate"
          },
          "artifact_digest": {
            "type": "string",
            "minLength": 1,
            "maxLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$"
          }
        },
        "required": [
          "schema_version",
          "job_id",
          "dataset_version",
          "config_version",
          "baseline_job_id",
          "completed_at",
          "manifest",
          "coverage",
          "metrics",
          "cost",
          "gate",
          "artifact_digest"
        ]
      },
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256,
            "pattern": "^urn:llm-regression-lab:error:[A-Z_]+$"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string",
            "enum": [
              "INVALID_JSON",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "IDEMPOTENCY_CONFLICT",
              "INVALID_STATE",
              "BASELINE_INCOMPATIBLE",
              "REPORT_NOT_READY",
              "REPORT_UNAVAILABLE",
              "ARTIFACT_EXPIRED",
              "PAYLOAD_TOO_LARGE",
              "VALIDATION_ERROR",
              "QUOTA_EXCEEDED",
              "INTERNAL_ERROR",
              "DEPENDENCY_UNAVAILABLE"
            ]
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          },
          "request_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "request_id",
          "retryable"
        ]
      },
...
    }
  }
}